+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Master
+
+## Bug Fixes
+
+- Initialising an Element with given meta or attributes as ObjectElement is now
+  supported.
+
 # 0.17.0 - 2017-06-16
 
 ## Breaking